+++ src/components/Order/Payment/index.tsx
@@ -1,10 +1,9 @@
-import React, { useState, useEffect, useMemo } from 'react';
 import Styled from '../styles';
 import InputField from '../InputField';
 import InputGroup from '../InputGroup';
 import PaymentList from '../List/PaymentList';
-import { getMemberCoupon, postMemberPoint } from 'src/apis/order/order';
-import { ICoupon } from 'src/types/order/types';
+import useDiscount from 'src/hooks/order/useDiscount';
+
 //상품 데이터
 const productData = {
   productId: 1,
@@ -16,73 +15,16 @@ const productData = {
 };
 
 const Payment = () => {
-  //쿠폰 적용
-  const [coupon, setCoupon] = useState<ICoupon[]>();
-  useEffect(() => {
-    getMemberCoupon().then(res => setCoupon(res));
-  }, []);
-  const couponOptions = useMemo(() => coupon, [coupon]);
-  //선택된 쿠폰
-  const [selectedCouponId, setSelectedCouponId] = useState<number>();
-  const [selectedCoupon, setSelectedCoupon] = useState<ICoupon>();
-  const [disabledCouponBtn, setDisabledCouponBtn] = useState(false);
-  const handleSelectedCoupon = event => {
-    const selectedValue = event.target.value;
-    setSelectedCouponId(Number(selectedValue));
-    //쿠폰을 선택하지 않았을 때
-    if (selectedValue === 0) {
-      setDisabledPointBtn(false);
-    }
-  };
-  useEffect(() => {
-    const selectedOption =
-      couponOptions &&
-      couponOptions?.find(coupon => coupon.couponId === selectedCouponId);
-    if (selectedOption) {
-      setSelectedCoupon(selectedOption);
-      setDisabledPointBtn(true);
-    } else {
-      setSelectedCoupon(null);
-      setDisabledPointBtn(false);
-    }
-  }, [selectedCouponId]);
-
-  // 적립금
-  const [point, setPoint] = useState<number>();
-  const [usedPoint, setUsedPoint] = useState<number>();
-  const [disabledPointBtn, setDisabledPointBtn] = useState(false);
-  useEffect(() => {
-    postMemberPoint().then(res => {
-      setPoint(res);
-      setUsedPoint(res);
-    });
-  }, []);
-
-  const handlePointChange = event => {
-    const typedPoint = Number(event.target.value);
-    setUsedPoint(typedPoint);
-    if (typedPoint > point) {
-      alert('최대로 사용할 수 있는 적립금을 초과하였습니다.');
-      setDisabledPointBtn(true);
-    }
-  };
-  // 할인된 가격
-  const [discountedPrice, setDiscountedPrice] = useState<number>();
-  console.log(selectedCoupon);
-  const getDiscountedPrice = (fullPrice: number) => {
-    if (selectedCoupon === null && usedPoint) {
-      setDiscountedPrice(usedPoint);
-      setDisabledCouponBtn(true);
-      return;
-    }
-    if (selectedCoupon && selectedCoupon.couponPolicy === 'FIXED') {
-      setDiscountedPrice(selectedCoupon.fixedPrice);
-    }
-    if (selectedCoupon && selectedCoupon.couponPolicy === 'RATE') {
-      const price = fullPrice * (selectedCoupon.rateAmount / 100);
-      setDiscountedPrice(price);
-    }
-  };
+  const {
+    coupon,
+    usedPoint,
+    handlePointChange,
+    handleSelectedCoupon,
+    disabledPointBtn,
+    disabledCouponBtn,
+    getDiscountedPrice,
+    discountedPrice,
+  } = useDiscount();
 
   return (
     <div>
@@ -98,7 +40,7 @@ const Payment = () => {
         <InputField
           label="쿠폰"
           field="coupon"
-          couponOptions={couponOptions}
+          couponOptions={coupon}
           handleSelectedCoupon={handleSelectedCoupon}
           disabledCouponBtn={disabledCouponBtn}
           getDiscountedPrice={() => getDiscountedPrice(productData.price)}
